refactor(experimental): move base58 code from rpc-core to transactions
## Summary

This will prevent a circular dependency in the next PR
