+++ packages/keys/package.json
@@ -64,7 +64,7 @@
         "node": ">=17.4"
     },
     "dependencies": {
-        "@metaplex-foundation/umi-serializers": "^0.8.2"
+        "@metaplex-foundation/umi-serializers": "^0.8.5"
     },
     "devDependencies": {
         "@solana/eslint-config-solana": "^1.0.2",
