+++ packages/rpc-core/package.json
@@ -61,12 +61,13 @@
         "maintained node versions"
     ],
     "dependencies": {
-        "@metaplex-foundation/umi-serializers-encodings": "^0.8.2",
-        "@solana/keys": "workspace:*"
+        "@metaplex-foundation/umi-serializers": "^0.8.5"
     },
     "devDependencies": {
         "@solana/eslint-config-solana": "^1.0.2",
+        "@solana/keys": "workspace:*",
         "@solana/rpc-transport": "workspace:*",
+        "@solana/transactions": "workspace:*",
         "@swc/core": "^1.3.18",
         "@swc/jest": "^0.2.26",
         "@types/jest": "^29.5.2",
