+++ packages/rpc-core/src/index.ts
@@ -1,4 +1,3 @@
-export * from './blockhash';
 export * from './rpc-methods';
 export * from './stringified-bigint';
 export * from './transaction-signature';
