+++ packages/transactions/package.json
@@ -61,14 +61,12 @@
         "maintained node versions"
     ],
     "dependencies": {
-        "@metaplex-foundation/umi-serializers": "^0.8.2",
+        "@metaplex-foundation/umi-serializers": "^0.8.5",
         "@solana/keys": "workspace:*"
     },
     "devDependencies": {
         "@solana/eslint-config-solana": "^1.0.2",
         "@solana/instructions": "workspace:*",
-        "@solana/keys": "workspace:*",
-        "@solana/rpc-core": "workspace:*",
         "@swc/core": "^1.3.18",
         "@swc/jest": "^0.2.23",
         "@types/jest": "^29.5.2",
